search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.1",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
